Split out `snoopi_deep` into separate, version-guarded files:
- src/parcel_snoopi_deep.jl
- SnoopCompileCore/src/snoopi_deep.jl
- test/snoopi_deep.jl
