+++ SnoopCompileCore/src/SnoopCompileCore.jl
@@ -11,6 +11,9 @@ if VERSION >= v"1.2.0-DEV.573"
     include("snoopi.jl")
 end
 
+if VERSION >= v"1.6.0-DEV.1190"  # https://github.com/JuliaLang/julia/pull/37749
+    include("snoopi_deep.jl")
+end
 
 if VERSION >= v"1.6.0-DEV.154"
     include("snoopr.jl")
